Modularize message create to a message controller
